Split account_entry from state mod. [skip travis]
